+++ lib/NativeReactor.php
@@ -242,7 +242,7 @@ class NativeReactor implements Reactor {
         $watcher->id = $watcherId = $this->lastWatcherId++;
         $watcher->type = Watcher::IMMEDIATE;
         $watcher->callback = $callback;
-        $watcher->callbackData = @$options["callback_data"];
+        $watcher->callbackData = @$options["cb_data"];
         $watcher->isEnabled = isset($options["enable"]) ? (bool) $options["enable"] : true;
 
         if ($watcher->isEnabled) {
@@ -268,7 +268,7 @@ class NativeReactor implements Reactor {
         $watcher->id = $watcherId = $this->lastWatcherId++;
         $watcher->type = Watcher::TIMER_ONCE;
         $watcher->callback = $callback;
-        $watcher->callbackData = @$options["callback_data"];
+        $watcher->callbackData = @$options["cb_data"];
         $watcher->isEnabled = isset($options["enable"]) ? (bool) $options["enable"] : true;
         $watcher->msDelay = round(($msDelay / 1000), 3);
         $watcher->nextExecutionAt = null;
@@ -303,7 +303,7 @@ class NativeReactor implements Reactor {
         $watcher->id = $watcherId = $this->lastWatcherId++;
         $watcher->type = Watcher::TIMER_REPEAT;
         $watcher->callback = $callback;
-        $watcher->callbackData = @$options["callback_data"];
+        $watcher->callbackData = @$options["cb_data"];
         $watcher->isEnabled = isset($options["enable"]) ? (bool) $options["enable"] : true;
         $watcher->msInterval = round(($msInterval / 1000), 3);
         $watcher->msDelay = round(($msDelay / 1000), 3);
@@ -345,7 +345,7 @@ class NativeReactor implements Reactor {
         $watcher->id = $watcherId = $this->lastWatcherId++;
         $watcher->type = $type;
         $watcher->callback = $callback;
-        $watcher->callbackData = @$options["callback_data"];
+        $watcher->callbackData = @$options["cb_data"];
         $watcher->isEnabled = isset($options["enable"]) ? (bool) $options["enable"] : true;
         $watcher->stream = $stream;
         $watcher->streamId = $streamId = (int) $stream;